+++ sdk/program/src/lib.rs
@@ -477,6 +477,7 @@ pub(crate) mod atomic_u64;
 pub mod big_mod_exp;
 pub mod blake3;
 pub mod borsh;
+pub mod borsh0_10;
 pub mod borsh0_9;
 pub mod bpf_loader;
 pub mod bpf_loader_deprecated;
